Update main navigation to be keyboard friendly.
Hover effects now also apply to focus events.

Sub-dropdown menus in the main navigation will now always be visibly hidden but focusable by keyboards and screen readers. Each individual link in a sub-dropdown menu starts with a width and height of 0 to visibly hide them. When they are focused (only doable via a keyboard or screen reader), the width and height are reset to automatic. This creates and effect that only shows one sub-dropdown link at a time when using a screen reader or keyboard. All sub-dropdown links are visible when the link iniating the sub-dropdown open is focused. Will need a nicer fix in the future.

If the user is using a mouse, there is no visible change.