Add new skip-to main content link above gold bar.
Create header-skipto HTML module and add it to the header HTML module.

Edit each HTML content module to replace the main content div with a <main> element with an id of "main".

Create a skipto scss file to contain styles for the skip-to main content link and apply to each main template scss file.