Restructure audience scss files.
Move the _audienceContent scss file into the audience folder.

Create a copy of the audience scss file and place it within the audience folder. Rename the file to audience-1.5.0 to label the version number. Modify the relative paths so the file will compile properly.

Keep the original audience scss file to support the original versions of the 2015 templates.