Restructure college scss files.
Move the _collegeContent scss file into the audience folder.

Create a copy of the college scss file and place it within the college folder. Rename the file to college-1.5.0 to label the version number. Modify the relative paths so the file will compile properly.

Keep the original college scss file to support the original versions of the 2015 templates.