Restructure content scss files.
Move the _contentPage and _sideNav scss files into the audience folder.

Create a copy of the content scss file and place it within the content folder. Rename the file to content-1.5.0 to label the version number. Modify the relative paths so the file will compile properly.

Keep the original content scss file to support the original versions of the 2015 templates.