Restructure deptOffice scss files.
Move the _deptOfficeContent scss file into the deptOffice folder.

Create a copy of the deptOffice scss file and place it within the deptOffice folder. Rename the file to deptOffice-1.5.0 to label the version number. Modify the relative paths so the file will compile properly.

Keep the original deptOffice scss file to support the original versions of the 2015 templates.