Restructure events scss files.
Move the _eventsContent scss file into the event folder.

Create a copy of the event scss file and place it within the vent folder. Rename the file to event-1.5.0 to label the version number. Modify the relative paths so the file will compile properly.

Keep the original event scss file to support the original versions of the 2015 templates.