Give default style to images and videos on content page.
When a .imageContainer or .vidContainer does not have a "right" or "left" class as well, then it will take up the full width. An image without a caption will take up the full width regardless of the width of the image. An image with a caption will take up the full width with it's limit being the images own width. A video will take the full width of the section.