Provide color and background-colors to account for no images.
The header was given a background-color of black to counteract the white text on white if the background-image fails to load. Apply the color white to the signature's link styles to counteract the black, alternate text being on the new black background if images fail to load.

The footer motto was given a font color of white for the alternate text being placed on a black background.

Each college tile is given a default, black background-color.

The proofpoints row (multiple images row) given white font color to account for alternate text on a dark-gray background.