Add extra examples of possible content for accordions.
Update the text of an accordion heading to change based on its index in the accordion group.

Change the name of accordion-list HTML module to accordion-content since it can be placed on more than lists.

Add arguments to accordion-content and accordion-item to change output based on its index in the accordion group (matching the accordion heading text).

Implement extra styles for accordions based on screensize.

Increase top and bottom padding of largest version of accordion.

Add styles to allow ordered lists that act as accordion content to output list item numbers.

Comment-out styles that change the background color of the accordion content and replace with a 15% gray background.

Comment-out styles that change the accordion item border colors and replace with black.