Add slim modifier to button and fix flexible padding on buttons.
Update the button container module to account for adding the "button-slim" class to a button if the slim parameter is equal to true.

Add a "Slim" section to the button pattern page.

Update the flexible padding on the button to allow the pseudo elements used as padding to shrink before any words within the button break to another line.